# PERF: StringEngine for string dtype indexing ops