--- a/pandas/_libs/index.pyx
+++ b/pandas/_libs/index.pyx
@@ -533,7 +533,7 @@ cdef class ObjectEngine(IndexEngine):
         return loc
 
 
-cdef class StringEngine(ObjectEngine):
+cdef class StringEngine(IndexEngine):
 
     cdef _make_hash_table(self, Py_ssize_t n):
         return _hash.StringHashTable(n)