… pandas/_libs/index.pyi